requirements.txt
aiodns==2.0.0
aiohttp==3.6.3
appdirs==1.4.4
arrow==0.17.0
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.2.0
beautifulsoup4==4.9.3
black==20.8b1
braceexpand==0.1.6
brotlipy==0.7.0
cchardet==2.1.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@6f2b873bf0d28903eb752aa1166b3aac26dc9007
discord.py==1.5.1
flake8==3.8.4
gTTS==2.1.1
gTTS-token==1.1.3
humanize==3.1.0
idna==2.10
import-expression==1.1.4
jishaku==1.19.2.202
langdetect==1.0.8
mccabe==0.6.1
multidict==4.7.6
mypy-extensions==0.4.3
pathspec==0.8.0
polyglot==16.7.4
pycares==3.1.1
# pycld2 from https://www.lfd.uci.edu/~gohlke/pythonlibs/#pycld2
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
# PyICU from https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyicu
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.28
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.7
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.5.1
